test: extract shared e2e harness helpers module
PR #72에서 SeaTable e2e harness 추가 시 ~250 LOC HELPERS 템플릿
리터럴이 4개 harness에 거의 동일하게 중복됐던 것을 정리.

추출 전략: harness가 CDP Runtime.evaluate로 plugin 컨텍스트에서
실행되므로 일반 ESM 함수 호출이 아니라 helper string을 합성해 HELPERS에
주입하는 패턴. tests/e2e/obsidian-helpers.mjs에 두 factory 함수 +
일반 ESM 유틸 두 개를 export:

- buildSyncHarnessHelpers({ pluginId, e2eCfgId }): sync harness용
  9 helper (getPlugin / getConfig / getCredential / getInstance /
  setMode / enqueueSync / waitForCache / openAndActivate / modifyField)
- buildSettingsHarnessHelpers({ pluginId }): settings harness용
  8 helper (getPlugin / getSettingsTab / openSettingsTab / rerenderTab /
  getContainer / queryCards / cardInfo / allCardInfos)
- makeSetConfigAndQuery({ helpers, run }): 두 settings harness가 공유
  하던 setConfigAndQuery 함수 본문을 factory로
- buildConfigExpr(overrides): cfg.<key> = <v> 라인 빌더 (일반 ESM)

부수 정합 작업:
- run-seatable-e2e.mjs: getSeaConfig/getSeaCredential → getConfig/
  getCredential rename (sync harness 명칭 통일, 16 호출 사이트 sed로 갱신)
- run-seatable-settings-e2e.mjs: getSeaConfig → getActiveConfig rename
  (두 settings harness가 같은 contract 함수명 공유 → setConfigAndQuery
  factory 통일 가능). fixture strategy(Airtable: 사용자 active 재사용 vs
  SeaTable: e2e cfg 주입 후 active set)는 setup 코드에 남아 있고 contract
  함수명만 통일됨.
- run-settings-e2e.mjs: 미사용 getConfig(idx) 제거

검증:
- node --check 5 파일 모두 통과
- vitest 519/519 PASS (e2e와 별개라 영향 없음)
- e2e cloud token이 필요한 npm run test:e2e / :seatable / :settings /
  :seatable:settings는 사용자 환경에서 별도 실행 필요

LOC delta:
- run-e2e.mjs: 1230 → 1160 (-70)
- run-seatable-e2e.mjs: 759 → 695 (-64)
- run-settings-e2e.mjs: 999 → 924 (-75)
- run-seatable-settings-e2e.mjs: 413 → 347 (-66)
- obsidian-helpers.mjs: +219
- 순 변화: 3401 → 3345 (-56 LOC)

순 감소는 작지만 4-way duplication을 1-way로 바꾼 게 본질이고,
다음 provider harness 추가 시 helper 재사용으로 80 LOC씩 절약 +
drift 위험 0.

Closes #74
